Moved coverage tests in a separate file
