+++ src/chanjo2/endpoints/coverage.py
@@ -1,6 +1,14 @@
+import logging
 from typing import List, Optional, Tuple
 
-from chanjo2.constants import WRONG_BED_FILE_MSG, WRONG_COVERAGE_FILE_MSG
+from fastapi import APIRouter, HTTPException, File, status, Depends
+from pyd4 import D4File
+from sqlmodel import Session
+
+from chanjo2.constants import WRONG_BED_FILE_MSG, WRONG_COVERAGE_FILE_MSG, SAMPLE_NOT_FOUND
+from chanjo2.crud.intervals import coordinates_from_interval_list
+from chanjo2.crud.samples import get_sample
+from chanjo2.dbutil import get_session
 from chanjo2.meta.handle_bed import parse_bed
 from chanjo2.meta.handle_d4 import (
     interval_coverage,
@@ -8,19 +16,19 @@ from chanjo2.meta.handle_d4 import (
     set_d4_file,
     set_interval,
 )
-from chanjo2.models.pydantic_models import CoverageInterval
-from fastapi import APIRouter, HTTPException, File, status
-from pyd4 import D4File
+from chanjo2.models.pydantic_models import CoverageInterval, SampleCoverageQuery
+from chanjo2.models.sql_models import Sample as SQLSample
 
+LOG = logging.getLogger("uvicorn.access")
 router = APIRouter()
 
 
 @router.get("/coverage/d4/interval/", response_model=CoverageInterval)
 def d4_interval_coverage(
-    coverage_file_path: str,
-    chromosome: str,
-    start: Optional[int] = None,
-    end: Optional[int] = None,
+        coverage_file_path: str,
+        chromosome: str,
+        start: Optional[int] = None,
+        end: Optional[int] = None,
 ):
     """Return coverage on the given interval for a D4 resource located on the disk or on a remote server."""
 
@@ -67,3 +75,26 @@ def d4_intervals_coverage(coverage_file_path: str, bed_file: bytes = File(...)):
         )
 
     return intervals_coverage(d4_file=d4_file, intervals=intervals)
+
+
+@router.post("/coverage/sample/interval_list", response_model=str)
+def sample_intervals_coverage(query: SampleCoverageQuery, db: Session = Depends(get_session)):
+    """Returns coverage over a list of intervals for a given sample in the database."""
+
+    sample: SQLSample = get_sample(db=db, sample_name=query.sample_name)
+    if sample is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=SAMPLE_NOT_FOUND,
+        )
+    try:
+        d4_file: D4File = set_d4_file(coverage_file_path=sample.coverage_file_path)
+    except Exception as ex:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=WRONG_COVERAGE_FILE_MSG,
+        )
+    coverage_intervals = coordinates_from_interval_list(intervals=query.intervals, interval_type=query.interval_type,
+                                                        build=query.build)
+
+    return f"You passed sample:{query.sample_name} - interval_list:{query.intervals} - interval_type:{query.interval_type}"
